benchmarking_tool/requirements.txt
alembic==1.4.3
APScheduler==3.6.3
Babel==2.9.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
boto3==1.16.24
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
demjson==2.2.4
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
flask-change-password==0.0.7
Flask-Login==0.5.0
Flask-Login-Multi==0.1.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Seeder==1.1.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
httplib2==0.18.1
idna==2.10
imutils==0.5.3
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jtutils==0.0.8
kiwisolver==1.3.1
leven==1.0.4
libcst==0.3.14
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mypy-extensions==0.4.3
nose==1.3.7
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
passlib==1.7.4
pdf2image==1.14.0
Pillow==8.0.1
proto-plus==1.11.0
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-csv==0.0.13
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
speaklater==1.3
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==1.2.0
xmltodict==0.12.0
